--- srcs/redir.c
@@ -1,3 +1,0 @@
-#include "../include/minishell.h"
-
-
