+++ srcs/syntax_checker.c
@@ -10,36 +10,36 @@ int		is_the_next_is_word(t_node *args)
 	return (0);
 }
 
-t_cmd   *print_syntax_error_cmd(t_node *args)
+t_cmd	*print_syntax_error_cmd(t_node *args)
 {
-    if (!args->next)
-    {
-        write(1, "[ERROR]: syntax error near unexpected token `", 45);
-        write(1, "newline", 7);
-        write(1, "'\n", 2);
-    }
-    else if (args->next->type != WORD)
-    {
-        write(1, "[ERROR]: syntax error near unexpected token `", 45);
-        write(1, args->next->value, ft_strlen(args->value));
-        write(1, "'\n", 2);
-    }
-    return (NULL);
+	if (!args->next)
+	{
+		write(1, "[ERROR]: syntax error near unexpected token `", 45);
+		write(1, "newline", 7);
+		write(1, "'\n", 2);
+	}
+	else if (args->next->type != WORD)
+	{
+		write(1, "[ERROR]: syntax error near unexpected token `", 45);
+		write(1, args->next->value, ft_strlen(args->value));
+		write(1, "'\n", 2);
+	}
+	return (NULL);
 }
 
-int   print_syntax_error_bool(t_node *args)
+int		print_syntax_error_bool(t_node *args)
 {
-    if (!args->next)
-    {
-        write(1, "[ERROR]: syntax error near unexpected token `", 45);
-        write(1, "newline", 7);
-        write(1, "'\n", 2);
-    }
-    else if (args->next->type != WORD)
-    {
-        write(1, "[ERROR]: syntax error near unexpected token `", 45);
-        write(1, args->next->value, ft_strlen(args->value));
-        write(1, "'\n", 2);
-    }
-    return (1);
+	if (!args->next)
+	{
+		write(1, "[ERROR]: syntax error near unexpected token `", 45);
+		write(1, "newline", 7);
+		write(1, "'\n", 2);
+	}
+	else if (args->next->type != WORD)
+	{
+		write(1, "[ERROR]: syntax error near unexpected token `", 45);
+		write(1, args->next->value, ft_strlen(args->value));
+		write(1, "'\n", 2);
+	}
+	return (1);
 }
